requirements.txt
Django==3.0.6
psycopg2==2.8.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.1
django-filter==2.2.0
requests==2.23.0